Switch to a platform-specific SizeType for container lengths.
The primarily avoids narrowing conversions from uint64_t to size_t
on 32bit platforms. The effect of this change is that 32bit
platforms generate early decode errors when handling containers
with element counts greater than 4GiB. This is unlikely to be a
problem given that 32bit platforms lack sufficient memory to
handle such sizes anyway.