Avoid bounds checks on std::vector::operator[] in debug builds.
Some implementations of std add bounds checks to operator[] in
debug builds, even though is it specified to not perform bounds
checks and it is valid to use the address of the element one past
the end of an array in interator contexts.

Use pointer arithmetic on the value of std::vector::data() instead.